requirements.txt
# Core ML/AI
# openai-whisper==20231117
torch>=2.0.0
transformers>=4.30.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20

# Vector DB and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Clustering
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.3

# Audio processing
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.0

# LLM integration
requests>=2.31.0

# UI
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
